Move monkeypatching relating to connections to classes
