+++ Gemfile
@@ -44,7 +44,7 @@ gem 'cancancan', '~> 1.10'
 gem 'carrierwave', '~> 1.3'
 gem 'phony_rails'
 gem 'responders'
-gem 'graphql', '< 1.11.0'
+gem 'graphql', '= 1.12.0'
 gem 'devise', '~> 4.8'
 gem 'devise-i18n', '~> 1.10.0'
 gem 'devise-two-factor', '~> 4.0.2'
