+++ app/graph/types/base_connection.rb
@@ -1,0 +1,3 @@
+class BaseConnection < GraphQL::Types::Relay::BaseConnection
+  field :totalCount, GraphQL::Types::Int, null: true
+end
