+++ app/graph/types/base_object.rb
@@ -1,3 +1,5 @@
 class BaseObject < GraphQL::Schema::Object
+  connection_type_class ::BaseConnection
+
   field_class SnakeCaseField
 end
