+++ app/graph/types/base_union.rb
@@ -1,2 +1,3 @@
 class BaseUnion < GraphQL::Schema::Union
+  connection_type_class ::BaseConnection
 end
