+++ app/lib/permissions_loader.rb
